+++ libs/feature-rooms/src/lib/message-composer/composer-toolbar/composer-toolbar.component.spec.ts
@@ -70,6 +70,7 @@ describe('ComposerToolbarComponent', () => {
       'format-codeblock',
       'format-quote',
       'format-list',
+      'format-tasklist',
     ]);
   });